Moved to XAppJwt and fixed namespaces
