+++ src/XUMM.NET.SDK/Models/XAppJwt/XummXAppJwtNFTokenDetail.cs
@@ -1,14 +1,17 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace XUMM.NET.SDK.Models.Misc;
+namespace XUMM.NET.SDK.Models.XAppJwt;
 
-public class XummNFTokenDetail
+public class XummXAppJwtNFTokenDetail
 {
     [JsonPropertyName("issuer")]
     public string? Issuer { get; set; }
 
     [JsonPropertyName("token")]
     public string? Token { get; set; }
+
+    [JsonPropertyName("owner")]
+    public string? Owner { get; set; }
 
     [JsonPropertyName("name")]
     public string? Name { get; set; }
