+++ src/XUMM.NET.SDK/Models/XAppJwt/XummXAppJwtUserDataResponse.cs
@@ -2,7 +2,7 @@
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace XUMM.NET.SDK.Models.XAppJWT;
+namespace XUMM.NET.SDK.Models.XAppJwt;
 
 public class XummXAppJwtUserDataResponse
 {
